requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
openai==1.99.9
supabase==2.28.3
yt-dlp==2026.3.17
youtube-transcript-api==1.2.4
pydantic==2.12.5
starlette==0.37.2
python-multipart==0.0.22
httpx==0.28.1
